notebooks added and tested with success
